Mental Health Rating vs Exam Scores Update
The mental_vs_exam script visualization functions. I decided to be unique and use a heat map. It came out well but not sure if it fits the best for this project. I'll keep it for now. Idk how to analyze the heatmap and come up with some results right now but I'll figure it out later. Next visualization I hope to make will be a histogram like the one I saw in chapter 1 of the dash reference.