Sleep Hours vs. Mental Health Rating runs
I fixed the sleep hours vs Mental Health Rating and turned it into a line plot. I'm either completely done with the bare visuals or almost done. We probably have to do the sleep vs exam scores visualization to make all the other visualizations make sense, so that's probably going to be the final visualization we make.